Implement listing of volumes and ISO download
Listing of volume is currently done through a combination of winrt and
win32 API calls. This works, but is a bit fragile. Furthermore, it makes
it hard to find the relationship between various volumes (e.g. if
they're on the same drive). Ideally, we should move to using the WMI
with the new Windows Storage Management API.

Downloading the ISO uses reqwest on a separate thread, and communicates
with the main thread through WizardEvents. This works-ish. It has
abyssal performance (should probably throttle the events to only happen
once every 500ms or something). A much better system would be to use
BackgroundDownloader from the WinRT APIs. Unfortunately, I was unable to
make it work. I believe it can only be done from an AppContainer (MSIX),
and not from a normal exe. I wonder if it's possible to generate a
temporary AppContainer from a .exe or something...